refactor(github): split GitHubService 2386-line monolith into focused modules
- Extract 10 module-scope caches into canonical GitHubCaches.ts leaf module
- Split domain logic into 9 focused modules following existing github/ pattern
- Replace GitHubService.ts with compatibility barrel preserving all exports
- Update getProjectHealth test mocks to target leaf modules

No behavior changes — pure mechanical refactor.
