+++ electron/services/GitHubService.ts
@@ -1,2386 +1,31 @@
-import type { GraphQlQueryResponseData } from "@octokit/graphql";
-import { GitService } from "./GitService.js";
-import { Cache } from "../utils/cache.js";
-import { GitHubStatsCache } from "./GitHubStatsCache.js";
-import { GitHubFirstPageCache } from "./GitHubFirstPageCache.js";
-import type {
-  GitHubIssue,
-  GitHubPR,
-  GitHubPRCIStatus,
-  GitHubPRCISummary,
-  GitHubUser,
-  GitHubListOptions,
-  GitHubListResponse,
-  LinkedPRInfo,
-  IssueTooltipData,
-  PRTooltipData,
-} from "../../shared/types/github.js";
-
-import {
-  GitHubAuth,
-  GITHUB_API_TIMEOUT_MS,
-  REPO_STATS_QUERY,
-  REPO_STATS_AND_PAGE_QUERY,
-  PROJECT_HEALTH_QUERY,
-  LIST_ISSUES_QUERY,
-  LIST_PRS_QUERY,
-  SEARCH_QUERY,
-  GET_ISSUE_QUERY,
-  GET_PR_QUERY,
-  buildBatchPRQuery,
-  buildBatchRequiredChecksQuery,
-  deriveRequiredCIStatus,
-  gitHubRateLimitService,
-  GitHubRateLimitError,
-} from "./github/index.js";
-import type { RollupContextNode } from "./github/index.js";
-import { getLastAuthMetadata } from "./github/GitHubAuth.js";
-import { formatErrorMessage } from "../../shared/utils/errorMessage.js";
-
-import type {
-  RepoContext,
-  RepoStats,
-  RepoStatsResult,
-  LinkedPR,
-  PRCheckResult,
-  PRCheckCandidate,
-  BatchPRCheckResult,
-  CIStatus,
-  ProjectHealth,
-  ProjectHealthResult,
+export {
+  getGitHubToken,
+  hasGitHubToken,
+  setGitHubToken,
+  clearGitHubToken,
+  getGitHubConfig,
+  getGitHubConfigAsync,
+  validateGitHubToken,
+  parseGitHubRepoUrl,
+  getRepoContext,
+  getRepoInfo,
+  getRepoUrl,
+  getIssueUrl,
+  withRepoContextRetry,
+  clearGitHubCaches,
+  clearPRCaches,
+  getRepoStats,
+  getRepoStatsAndPage,
+  getProjectHealth,
+  batchCheckLinkedPRs,
+  parseGitHubError,
+  listIssues,
+  listPullRequests,
+  assignIssue,
+  getIssueTooltip,
+  getPRTooltip,
+  getIssueByNumber,
+  getPRByNumber,
 } from "./github/index.js";
 
-export type { GitHubTokenConfig, GitHubTokenValidation } from "./github/index.js";
-export type {
-  RepoContext,
-  RepoStats,
-  RepoStatsResult,
-  LinkedPR,
-  PRCheckResult,
-  PRCheckCandidate,
-  BatchPRCheckResult,
-  CIStatus,
-  ProjectHealth,
-  ProjectHealthResult,
-};
-
-// Caches
-const repoContextCache = new Cache<string, RepoContext>({ defaultTTL: 300000 });
-const repoStatsCache = new Cache<string, RepoStats>({ defaultTTL: 60000 });
-const issueListCache = new Cache<string, GitHubListResponse<GitHubIssue>>({ defaultTTL: 60000 });
-const prListCache = new Cache<string, GitHubListResponse<GitHubPR>>({ defaultTTL: 60000 });
-const projectHealthCache = new Cache<string, ProjectHealth>({ defaultTTL: 60000 });
-const issueTooltipCache = new Cache<string, IssueTooltipData>({ defaultTTL: 300000 }); // 5 min TTL
-// Newest-wins guard for prTooltipCache: tracks the request-start timestamp of
-// the most recent successful write per cache key. Both poll-path (pre-warm)
-// and hover-path writes capture Date.now() before awaiting the network call,
-// then only commit when their captured timestamp is >= the stored one. Without
-// this guard a slow poll resolving after a faster hover fetch would clobber
-// fresher data with staler data (lessons #2243, #1377). The Cache `onEvict`
-// hook keeps this map's lifetime tied to prTooltipCache so LRU eviction does
-// not leave orphaned timestamps; explicit clears below also reset it.
-const prTooltipWrittenAt = new Map<string, number>();
-const prTooltipCache = new Cache<string, PRTooltipData>({
-  defaultTTL: 300000, // 5 min TTL
-  onEvict: (key) => {
-    prTooltipWrittenAt.delete(key as string);
-  },
-});
-
-// ETag cache for PR conditional revalidation. Key: `owner/repo#prNumber`.
-// Cleared on token rotation via clearGitHubCaches (ETags are token-scoped).
-const prETagCache = new Map<string, string>();
-
-// ETag cache for branch-to-PR list conditional probes used by the discovery
-// path. Key: `owner/repo@branchName`. Cleared atomically with `prETagCache`
-// in `clearGitHubCaches` because ETags are token-scoped.
-const branchListETagCache = new Map<string, string>();
-
-interface PRRequiredStatusEntry {
-  ciStatus: GitHubPRCIStatus | undefined;
-  ciSummary: GitHubPRCISummary | undefined;
-}
-// Key: `${owner}/${repo}:${prNumber}`
-const prRequiredStatusCache = new Cache<string, PRRequiredStatusEntry>({ defaultTTL: 60000 });
-
-export function getGitHubToken(): string | undefined {
-  return GitHubAuth.getToken();
-}
-
-export function hasGitHubToken(): boolean {
-  return GitHubAuth.hasToken();
-}
-
-export function setGitHubToken(token: string): void {
-  GitHubAuth.setToken(token);
-  clearGitHubCaches();
-}
-
-export function clearGitHubToken(): void {
-  GitHubAuth.clearToken();
-  clearGitHubCaches();
-}
-
-export function getGitHubConfig() {
-  return GitHubAuth.getConfig();
-}
-
-export async function getGitHubConfigAsync() {
-  return GitHubAuth.getConfigAsync();
-}
-
-export async function validateGitHubToken(token: string) {
-  return GitHubAuth.validate(token);
-}
-
-/**
- * Parse owner and repo from a GitHub URL (HTTPS or SSH format).
- *
- * Handles:
- * - HTTPS URLs: https://github.com/owner/repo, https://github.com/owner/repo.git
- * - SSH URLs: git@github.com:owner/repo, git@github.com:owner/repo.git
- * - Trailing slashes: https://github.com/owner/repo/
- * - Dotted repo names: https://github.com/owner/my.repo
- */
-export function parseGitHubRepoUrl(url: string): { owner: string; repo: string } | null {
-  // Normalize SSH format to HTTPS-like format for easier parsing
-  const normalized = url.replace(/^git@github\.com:/, "https://github.com/");
-
-  try {
-    const parsed = new URL(normalized);
-
-    // Must be github.com
-    if (parsed.hostname !== "github.com") {
-      return null;
-    }
-
-    // Remove leading slash, trailing slash, and .git suffix
-    const pathname = parsed.pathname
-      .replace(/^\//, "")
-      .replace(/\/$/, "")
-      .replace(/\.git$/, "");
-
-    const parts = pathname.split("/");
-
-    // Need at least owner and repo
-    if (parts.length < 2 || !parts[0] || !parts[1]) {
-      return null;
-    }
-
-    return { owner: parts[0], repo: parts[1] };
-  } catch {
-    return null;
-  }
-}
-
-export async function getRepoContext(cwd: string): Promise<RepoContext | null> {
-  const cached = repoContextCache.get(cwd);
-  if (cached) return cached;
-
-  try {
-    const gitService = new GitService(cwd);
-
-    let fetchUrl: string | null = null;
-
-    try {
-      const { projectStore } = await import("./ProjectStore.js");
-      const repoRoot = await gitService.getRepositoryRoot(cwd);
-      const project = await projectStore.getProjectByPath(repoRoot);
-      if (project) {
-        const settings = await projectStore.getProjectSettings(project.id);
-        if (settings.githubRemote) {
-          const remotes = await gitService.listRemotes(cwd);
-          const match = remotes.find((r) => r.name === settings.githubRemote);
-          if (match?.fetchUrl) {
-            fetchUrl = match.fetchUrl;
-          }
-        }
-      }
-    } catch {
-      // Fall through to default origin lookup
-    }
-
-    if (!fetchUrl) {
-      fetchUrl = await gitService.getRemoteUrl(cwd);
-    }
-
-    if (!fetchUrl) return null;
-
-    const parsed = parseGitHubRepoUrl(fetchUrl);
-    if (!parsed) return null;
-
-    const context = { owner: parsed.owner, repo: parsed.repo };
-    repoContextCache.set(cwd, context);
-    return context;
-  } catch {
-    return null;
-  }
-}
-
-function isRepoNotFoundError(error: unknown): boolean {
-  const message = formatErrorMessage(error, "Failed to access GitHub repository");
-  const lower = message.toLowerCase();
-  // Broad pattern match - includes both repo-level and resource-level "not found" errors.
-  // This is acceptable because withRepoContextRetry only retries when repo context actually
-  // changes after invalidation, preventing false retries on genuine resource 404s.
-  return lower.includes("not found") || lower.includes("could not resolve");
-}
-
-async function withRepoContextRetry<T>(
-  cwd: string,
-  fn: (context: RepoContext) => Promise<T>
-): Promise<T> {
-  const context = await getRepoContext(cwd);
-  if (!context) throw new Error("Not a GitHub repository");
-
-  try {
-    return await fn(context);
-  } catch (error) {
-    if (isRepoNotFoundError(error)) {
-      repoContextCache.invalidate(cwd);
-      const freshContext = await getRepoContext(cwd);
-      if (
-        freshContext &&
-        (freshContext.owner !== context.owner || freshContext.repo !== context.repo)
-      ) {
-        return await fn(freshContext);
-      }
-    }
-    throw error;
-  }
-}
-
-export async function getRepoStats(
-  cwd: string,
-  bypassCache = false,
-  _retried = false
-): Promise<RepoStatsResult> {
-  const context = await getRepoContext(cwd);
-  if (!context) {
-    return { stats: null, error: "Not a GitHub repository" };
-  }
-
-  const cacheKey = `${context.owner}/${context.repo}`;
-  const persistentCache = GitHubStatsCache.getInstance();
-
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    const diskCached = persistentCache.get(cacheKey);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        error: "GitHub token not configured. Set it in Settings.",
-      };
-    }
-    return { stats: null, error: "GitHub token not configured. Set it in Settings." };
-  }
-
-  const rateLimitBlock = gitHubRateLimitService.shouldBlockRequest();
-  if (rateLimitBlock.blocked && rateLimitBlock.reason && rateLimitBlock.resumeAt) {
-    const diskCached = persistentCache.get(cacheKey);
-    const message = rateLimitMessage(rateLimitBlock.reason, rateLimitBlock.resumeAt);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        error: message,
-      };
-    }
-    return { stats: null, error: message };
-  }
-
-  if (!bypassCache) {
-    const cached = repoStatsCache.get(cacheKey);
-    if (cached) {
-      return { stats: cached };
-    }
-  }
-
-  try {
-    const result = (await client(REPO_STATS_QUERY, {
-      owner: context.owner,
-      repo: context.repo,
-      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-    })) as GraphQlQueryResponseData;
-
-    const repository = result?.repository;
-    if (!repository) {
-      const diskCached = persistentCache.get(cacheKey);
-      if (diskCached) {
-        return {
-          stats: {
-            issueCount: diskCached.issueCount,
-            prCount: diskCached.prCount,
-            stale: true,
-            lastUpdated: diskCached.lastUpdated,
-          },
-          error: "Repository not found (showing cached data)",
-        };
-      }
-      return { stats: null, error: "Repository not found" };
-    }
-
-    const stats: RepoStats = {
-      issueCount: repository.issues?.totalCount ?? 0,
-      prCount: repository.pullRequests?.totalCount ?? 0,
-      lastUpdated: Date.now(),
-    };
-
-    repoStatsCache.set(cacheKey, stats);
-    persistentCache.set(cacheKey, stats, cwd);
-
-    return { stats };
-  } catch (error) {
-    if (!_retried && isRepoNotFoundError(error)) {
-      repoContextCache.invalidate(cwd);
-      const freshContext = await getRepoContext(cwd);
-      if (
-        freshContext &&
-        (freshContext.owner !== context.owner || freshContext.repo !== context.repo)
-      ) {
-        return getRepoStats(cwd, bypassCache, true);
-      }
-    }
-    const diskCached = persistentCache.get(cacheKey);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        error: parseGitHubError(error),
-      };
-    }
-    return { stats: null, error: parseGitHubError(error) };
-  }
-}
-
-export interface RepoStatsAndPageResult {
-  stats: RepoStats | null;
-  issues: {
-    items: GitHubIssue[];
-    endCursor: string | null;
-    hasNextPage: boolean;
-    totalCount: number;
-  } | null;
-  prs: {
-    items: GitHubPR[];
-    endCursor: string | null;
-    hasNextPage: boolean;
-    totalCount: number;
-  } | null;
-  /**
-   * `"network"` when the result came from a fresh GraphQL round-trip;
-   * `"memory-cache"` when the function short-circuited against the local
-   * in-memory caches. Callers (the IPC handler) use this to avoid
-   * broadcasting cache-extended data that would refresh the renderer's TTL
-   * with no genuine freshness gain.
-   */
-  source?: "network" | "memory-cache";
-  error?: string;
-}
-
-/**
- * Combined fetch: counts + first page of open issues + open PRs in a single
- * GraphQL round-trip. Used by the toolbar poll so the renderer can prime its
- * `githubResourceCache` for the (open, created) default-filter cache key with
- * no click-time round-trip. Falls back to disk-cached counts on auth/rate
- * issues exactly like `getRepoStats` does, returning `issues: null` / `prs:
- * null` so callers can choose to broadcast counts only.
- */
-export async function getRepoStatsAndPage(
-  cwd: string,
-  bypassCache = false,
-  _retried = false
-): Promise<RepoStatsAndPageResult> {
-  const context = await getRepoContext(cwd);
-  if (!context) {
-    return { stats: null, issues: null, prs: null, error: "Not a GitHub repository" };
-  }
-
-  const cacheKey = `${context.owner}/${context.repo}`;
-  const persistentCache = GitHubStatsCache.getInstance();
-  const client = GitHubAuth.createClient();
-
-  if (!client) {
-    const diskCached = persistentCache.get(cacheKey);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        issues: null,
-        prs: null,
-        error: "GitHub token not configured. Set it in Settings.",
-      };
-    }
-    return {
-      stats: null,
-      issues: null,
-      prs: null,
-      error: "GitHub token not configured. Set it in Settings.",
-    };
-  }
-
-  const rateLimitBlock = gitHubRateLimitService.shouldBlockRequest();
-  if (rateLimitBlock.blocked && rateLimitBlock.reason && rateLimitBlock.resumeAt) {
-    const diskCached = persistentCache.get(cacheKey);
-    const message = rateLimitMessage(rateLimitBlock.reason, rateLimitBlock.resumeAt);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        issues: null,
-        prs: null,
-        error: message,
-      };
-    }
-    return { stats: null, issues: null, prs: null, error: message };
-  }
-
-  // Memory-cache short-circuit: if counts cache is hot AND we already have a
-  // primed `issueListCache` / `prListCache` for the default filter, skip the
-  // full round-trip and return what we already have. The list caches share
-  // the 60s TTL, so when one is hot the other usually is too.
-  if (!bypassCache) {
-    const cachedStats = repoStatsCache.get(cacheKey);
-    const issuesCacheKey = buildListCacheKey(
-      "issue",
-      context.owner,
-      context.repo,
-      "open",
-      "",
-      "created",
-      ""
-    );
-    const prsCacheKey = buildListCacheKey(
-      "pr",
-      context.owner,
-      context.repo,
-      "open",
-      "",
-      "created",
-      ""
-    );
-    const cachedIssues = issueListCache.get(issuesCacheKey);
-    const cachedPRs = prListCache.get(prsCacheKey);
-    if (cachedStats && cachedIssues && cachedPRs) {
-      return {
-        stats: cachedStats,
-        issues: {
-          items: cachedIssues.items,
-          endCursor: cachedIssues.pageInfo.endCursor,
-          hasNextPage: cachedIssues.pageInfo.hasNextPage,
-          totalCount: cachedStats.issueCount,
-        },
-        prs: {
-          items: cachedPRs.items,
-          endCursor: cachedPRs.pageInfo.endCursor,
-          hasNextPage: cachedPRs.pageInfo.hasNextPage,
-          totalCount: cachedStats.prCount,
-        },
-        source: "memory-cache",
-      };
-    }
-  }
-
-  try {
-    const result = (await client(REPO_STATS_AND_PAGE_QUERY, {
-      owner: context.owner,
-      repo: context.repo,
-      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-    })) as GraphQlQueryResponseData;
-
-    const repository = result?.repository;
-    if (!repository) {
-      const diskCached = persistentCache.get(cacheKey);
-      if (diskCached) {
-        return {
-          stats: {
-            issueCount: diskCached.issueCount,
-            prCount: diskCached.prCount,
-            stale: true,
-            lastUpdated: diskCached.lastUpdated,
-          },
-          issues: null,
-          prs: null,
-          error: "Repository not found (showing cached data)",
-        };
-      }
-      return { stats: null, issues: null, prs: null, error: "Repository not found" };
-    }
-
-    const issuesData = repository.issues as
-      | {
-          totalCount?: number;
-          pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-          nodes?: Array<Record<string, unknown>>;
-        }
-      | undefined;
-    const prsData = repository.pullRequests as
-      | {
-          totalCount?: number;
-          pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-          nodes?: Array<Record<string, unknown>>;
-        }
-      | undefined;
-
-    const issueCount = issuesData?.totalCount ?? 0;
-    const prCount = prsData?.totalCount ?? 0;
-    const stats: RepoStats = {
-      issueCount,
-      prCount,
-      lastUpdated: Date.now(),
-    };
-
-    repoStatsCache.set(cacheKey, stats);
-    persistentCache.set(cacheKey, stats, cwd);
-
-    const parsedIssues = (issuesData?.nodes ?? []).filter(Boolean).map(parseIssueNode);
-    const parsedPRs = (prsData?.nodes ?? []).filter(Boolean).map(parsePRNode);
-
-    // Persist the first page to disk so that on the next cold start we can
-    // hydrate the renderer's resource cache before any network poll completes.
-    GitHubFirstPageCache.getInstance().set(
-      cacheKey,
-      {
-        issues: {
-          items: parsedIssues,
-          endCursor: issuesData?.pageInfo?.endCursor ?? null,
-          hasNextPage: issuesData?.pageInfo?.hasNextPage ?? false,
-        },
-        prs: {
-          items: parsedPRs,
-          endCursor: prsData?.pageInfo?.endCursor ?? null,
-          hasNextPage: prsData?.pageInfo?.hasNextPage ?? false,
-        },
-      },
-      cwd
-    );
-
-    // Prime the per-list memory caches under the default-filter key so a
-    // subsequent click-time `listIssues({ state: "open", sortOrder:
-    // "created" })` resolves from cache.
-    const issuesListCacheKey = buildListCacheKey(
-      "issue",
-      context.owner,
-      context.repo,
-      "open",
-      "",
-      "created",
-      ""
-    );
-    const prsListCacheKey = buildListCacheKey(
-      "pr",
-      context.owner,
-      context.repo,
-      "open",
-      "",
-      "created",
-      ""
-    );
-    const issuesPage = {
-      items: parsedIssues,
-      endCursor: issuesData?.pageInfo?.endCursor ?? null,
-      hasNextPage: issuesData?.pageInfo?.hasNextPage ?? false,
-      totalCount: issueCount,
-    };
-    const prsPage = {
-      items: parsedPRs,
-      endCursor: prsData?.pageInfo?.endCursor ?? null,
-      hasNextPage: prsData?.pageInfo?.hasNextPage ?? false,
-      totalCount: prCount,
-    };
-    issueListCache.set(issuesListCacheKey, {
-      items: issuesPage.items,
-      pageInfo: { hasNextPage: issuesPage.hasNextPage, endCursor: issuesPage.endCursor },
-    });
-    prListCache.set(prsListCacheKey, {
-      items: prsPage.items,
-      pageInfo: { hasNextPage: prsPage.hasNextPage, endCursor: prsPage.endCursor },
-    });
-
-    return { stats, issues: issuesPage, prs: prsPage, source: "network" };
-  } catch (error) {
-    if (!_retried && isRepoNotFoundError(error)) {
-      repoContextCache.invalidate(cwd);
-      const freshContext = await getRepoContext(cwd);
-      if (
-        freshContext &&
-        (freshContext.owner !== context.owner || freshContext.repo !== context.repo)
-      ) {
-        return getRepoStatsAndPage(cwd, bypassCache, true);
-      }
-    }
-    const diskCached = persistentCache.get(cacheKey);
-    if (diskCached) {
-      return {
-        stats: {
-          issueCount: diskCached.issueCount,
-          prCount: diskCached.prCount,
-          stale: true,
-          lastUpdated: diskCached.lastUpdated,
-        },
-        issues: null,
-        prs: null,
-        error: parseGitHubError(error),
-      };
-    }
-    return { stats: null, issues: null, prs: null, error: parseGitHubError(error) };
-  }
-}
-
-function parseCIStatus(statusCheckRollup: { state?: string } | null | undefined): CIStatus {
-  if (!statusCheckRollup) return "none";
-  const state = statusCheckRollup.state?.toUpperCase();
-  switch (state) {
-    case "SUCCESS":
-      return "success";
-    case "FAILURE":
-      return "failure";
-    case "ERROR":
-      return "error";
-    case "PENDING":
-      return "pending";
-    case "EXPECTED":
-      return "expected";
-    default:
-      return "none";
-  }
-}
-
-function extractMergedCounts(data: Record<string, unknown>): Record<60 | 120 | 180, number> {
-  const getCount = (key: string): number => {
-    const entry = data[key] as { issueCount?: number } | undefined;
-    return entry?.issueCount ?? 0;
-  };
-  return {
-    60: getCount("mergedPRs60"),
-    120: getCount("mergedPRs120"),
-    180: getCount("mergedPRs180"),
-  };
-}
-
-function parseProjectHealthResponse(
-  repository: Record<string, unknown>,
-  mergedCounts: Record<60 | 120 | 180, number>,
-  repoUrl: string
-): ProjectHealth {
-  const defaultBranchRef = repository.defaultBranchRef as {
-    target?: { statusCheckRollup?: { state?: string } | null };
-  } | null;
-  const statusCheckRollup = defaultBranchRef?.target?.statusCheckRollup ?? null;
-
-  const latestRelease = repository.latestRelease as {
-    tagName: string;
-    publishedAt: string | null;
-    url: string;
-  } | null;
-
-  const vulnerabilityAlerts = repository.vulnerabilityAlerts as {
-    totalCount: number;
-  } | null;
-
-  return {
-    ciStatus: parseCIStatus(statusCheckRollup),
-    issueCount: (repository.issues as { totalCount?: number })?.totalCount ?? 0,
-    prCount: (repository.pullRequests as { totalCount?: number })?.totalCount ?? 0,
-    latestRelease: latestRelease
-      ? {
-          tagName: latestRelease.tagName,
-          publishedAt: latestRelease.publishedAt,
-          url: latestRelease.url,
-        }
-      : null,
-    securityAlerts: {
-      visible: vulnerabilityAlerts != null,
-      count: vulnerabilityAlerts?.totalCount ?? 0,
-    },
-    mergeVelocity: {
-      mergedCounts,
-    },
-    repoUrl,
-    lastUpdated: Date.now(),
-  };
-}
-
-export async function getProjectHealth(
-  cwd: string,
-  bypassCache = false,
-  _retried = false
-): Promise<ProjectHealthResult> {
-  const context = await getRepoContext(cwd);
-  if (!context) {
-    return { health: null, error: "Not a GitHub repository" };
-  }
-
-  const cacheKey = `${context.owner}/${context.repo}`;
-  const repoUrl = `https://github.com/${context.owner}/${context.repo}`;
-
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    return { health: null, error: "GitHub token not configured. Set it in Settings." };
-  }
-
-  const rateLimitBlock = gitHubRateLimitService.shouldBlockRequest();
-  if (rateLimitBlock.blocked && rateLimitBlock.reason && rateLimitBlock.resumeAt) {
-    return {
-      health: null,
-      error: rateLimitMessage(rateLimitBlock.reason, rateLimitBlock.resumeAt),
-    };
-  }
-
-  if (!bypassCache) {
-    const cached = projectHealthCache.get(cacheKey);
-    if (cached) {
-      return { health: cached };
-    }
-  }
-
-  try {
-    const repoQualifier = `repo:${context.owner}/${context.repo}`;
-    const mergedSearchBase = `${repoQualifier} is:pr is:merged`;
-    const now = new Date();
-    const mergedQueryVars = Object.fromEntries(
-      ([60, 120, 180] as const).map((days) => {
-        const since = new Date(now.getTime() - days * 24 * 60 * 60 * 1000);
-        const dateStr = since.toISOString().slice(0, 10);
-        return [`merged${days}`, `${mergedSearchBase} merged:>=${dateStr}`];
-      })
-    );
-
-    const result = (await client(PROJECT_HEALTH_QUERY, {
-      owner: context.owner,
-      repo: context.repo,
-      ...mergedQueryVars,
-      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-    })) as GraphQlQueryResponseData;
-
-    const repository = result?.repository;
-    if (!repository) {
-      return { health: null, error: "Repository not found" };
-    }
-
-    const mergedCounts = extractMergedCounts(result);
-    const health = parseProjectHealthResponse(
-      repository as Record<string, unknown>,
-      mergedCounts,
-      repoUrl
-    );
-    projectHealthCache.set(cacheKey, health);
-    return { health };
-  } catch (error: unknown) {
-    // @octokit/graphql throws GraphqlResponseError on partial failures
-    // (e.g., vulnerabilityAlerts FORBIDDEN). Extract partial data if available.
-    if (error instanceof Error && "data" in error && (error as { data?: unknown }).data) {
-      const partialData = (error as { data: Record<string, unknown> }).data;
-      const repository = partialData?.repository as Record<string, unknown> | undefined;
-      if (repository) {
-        const mergedCounts = extractMergedCounts(partialData);
-        const health = parseProjectHealthResponse(repository, mergedCounts, repoUrl);
-        projectHealthCache.set(cacheKey, health);
-        return { health };
-      }
-    }
-
-    if (!_retried && isRepoNotFoundError(error)) {
-      repoContextCache.invalidate(cwd);
-      const freshContext = await getRepoContext(cwd);
-      if (
-        freshContext &&
-        (freshContext.owner !== context.owner || freshContext.repo !== context.repo)
-      ) {
-        return getProjectHealth(cwd, bypassCache, true);
-      }
-    }
-
-    return { health: null, error: parseGitHubError(error) };
-  }
-}
-
-export async function getRepoInfo(cwd: string): Promise<RepoContext | null> {
-  return getRepoContext(cwd);
-}
-
-/**
- * Tooltip-shaped raw fields fetched alongside `LinkedPR` shape in the batch
- * query. Mirrors the subset of `GET_PR_QUERY` fields that `getPRTooltip`
- * reads, so the same projection logic produces a `PRTooltipData`.
- */
-interface BatchPRTooltipFields {
-  bodyText?: string | null;
-  createdAt?: string;
-  author?: { login?: string; avatarUrl?: string } | null;
-  assignees?: { nodes?: Array<{ login?: string; avatarUrl?: string }> };
-  labels?: { nodes?: Array<{ name?: string; color?: string }> };
-}
-
-interface BatchPRRawNode extends BatchPRTooltipFields {
-  number?: number;
-  title?: string;
-  url?: string;
-  state?: string;
-  isDraft?: boolean;
-  merged?: boolean;
-}
-
-/**
- * Build a `PRTooltipData` from the raw GraphQL node returned by the batch
- * query. Mirrors the projection in `getPRTooltip` (single-PR path) so both
- * write paths produce structurally identical cache entries. Returns
- * `undefined` when required fields are missing — the batch query may degrade
- * gracefully if a node was selected via the `LinkedPR` shape only (e.g., on
- * a partial response).
- */
-function buildTooltipDataFromBatchNode(node: BatchPRRawNode): PRTooltipData | undefined {
-  if (typeof node.number !== "number" || typeof node.title !== "string") {
-    return undefined;
-  }
-  if (typeof node.createdAt !== "string") {
-    // createdAt is a required scalar on PRTooltipData — without it we cannot
-    // construct a valid entry (callers format it as a date).
-    return undefined;
-  }
-
-  const merged = node.merged ?? false;
-  const rawState = (node.state ?? "OPEN").toUpperCase();
-  const state: "OPEN" | "CLOSED" | "MERGED" = merged
-    ? "MERGED"
-    : rawState === "CLOSED"
-      ? "CLOSED"
-      : "OPEN";
-
-  return {
-    number: node.number,
-    title: node.title,
-    bodyExcerpt: truncateBody(node.bodyText ?? null),
-    state,
-    isDraft: node.isDraft ?? false,
-    createdAt: node.createdAt,
-    author: {
-      login: node.author?.login ?? "unknown",
-      avatarUrl: node.author?.avatarUrl ?? "",
-    },
-    assignees: (node.assignees?.nodes ?? []).filter(Boolean).map((a) => ({
-      login: a.login ?? "unknown",
-      avatarUrl: a.avatarUrl ?? "",
-    })),
-    labels: (node.labels?.nodes ?? []).filter(Boolean).map((l) => ({
-      name: l.name ?? "",
-      color: l.color ?? "",
-    })),
-  };
-}
-
-function parseBatchPRResponse(
-  data: Record<string, unknown>,
-  candidates: PRCheckCandidate[]
-): Map<string, PRCheckResult> {
-  const results = new Map<string, PRCheckResult>();
-
-  for (let i = 0; i < candidates.length; i++) {
-    const candidate = candidates[i];
-    const alias = `wt_${i}`;
-    let foundPR: LinkedPR | null = null;
-    let foundTooltip: PRTooltipData | undefined;
-    const issueResponse = (
-      data?.[`${alias}_issue`] as {
-        issue?: { title?: string; timelineItems?: { nodes?: unknown[] } };
-      }
-    )?.issue;
-    const issueTitle = issueResponse?.title;
-    const issueData = issueResponse?.timelineItems?.nodes;
-    if (issueData && Array.isArray(issueData)) {
-      const prs: Array<{ pr: LinkedPR; raw: BatchPRRawNode }> = [];
-      const seenPRNumbers = new Set<number>();
-      for (const node of issueData as Array<{
-        source?: BatchPRRawNode;
-        subject?: BatchPRRawNode;
-      }>) {
-        // Handle both CROSS_REFERENCED_EVENT (source) and CONNECTED_EVENT (subject)
-        const prData = node?.source ?? node?.subject;
-        if (prData?.number && prData?.url && !seenPRNumbers.has(prData.number)) {
-          seenPRNumbers.add(prData.number);
-          prs.push({
-            pr: {
-              number: prData.number,
-              title: prData.title || "",
-              url: prData.url,
-              state: prData.merged
-                ? "merged"
-                : (prData.state?.toLowerCase() as "open" | "closed") || "open",
-              isDraft: prData.isDraft ?? false,
-            },
-            raw: prData,
-          });
-        }
-      }
-
-      const openPRs = prs.filter((entry) => entry.pr.state === "open");
-      const mergedPRs = prs.filter((entry) => entry.pr.state === "merged");
-      const closedPRs = prs.filter((entry) => entry.pr.state === "closed");
-
-      let chosen: { pr: LinkedPR; raw: BatchPRRawNode } | undefined;
-      if (openPRs.length > 0) {
-        chosen = openPRs[openPRs.length - 1];
-      } else if (mergedPRs.length > 0) {
-        chosen = mergedPRs[mergedPRs.length - 1];
-      } else if (closedPRs.length > 0) {
-        chosen = closedPRs[closedPRs.length - 1];
-      }
-      if (chosen) {
-        foundPR = chosen.pr;
-        foundTooltip = buildTooltipDataFromBatchNode(chosen.raw);
-      }
-    }
-
-    if (!foundPR) {
-      const branchData = (data?.[`${alias}_branch`] as { pullRequests?: { nodes?: unknown[] } })
-        ?.pullRequests?.nodes;
-      if (branchData && Array.isArray(branchData)) {
-        // Parse all PRs from branch query
-        const branchPRs: Array<{ pr: LinkedPR; raw: BatchPRRawNode }> = [];
-        for (const node of branchData as BatchPRRawNode[]) {
-          if (node?.number && node?.url) {
-            branchPRs.push({
-              pr: {
-                number: node.number,
-                title: node.title || "",
-                url: node.url,
-                state: node.merged
-                  ? "merged"
-                  : (node.state?.toLowerCase() as "open" | "closed") || "open",
-                isDraft: node.isDraft ?? false,
-              },
-              raw: node,
-            });
-          }
-        }
-
-        // Apply same preference: open > merged > closed
-        const openPRs = branchPRs.filter((entry) => entry.pr.state === "open");
-        const mergedPRs = branchPRs.filter((entry) => entry.pr.state === "merged");
-        const closedPRs = branchPRs.filter((entry) => entry.pr.state === "closed");
-
-        let chosen: { pr: LinkedPR; raw: BatchPRRawNode } | undefined;
-        if (openPRs.length > 0) {
-          chosen = openPRs[openPRs.length - 1];
-        } else if (mergedPRs.length > 0) {
-          chosen = mergedPRs[mergedPRs.length - 1];
-        } else if (closedPRs.length > 0) {
-          chosen = closedPRs[closedPRs.length - 1];
-        }
-        if (chosen) {
-          foundPR = chosen.pr;
-          foundTooltip = buildTooltipDataFromBatchNode(chosen.raw);
-        }
-      }
-    }
-
-    results.set(candidate.worktreeId, {
-      issueNumber: candidate.issueNumber,
-      issueTitle,
-      branchName: candidate.branchName,
-      pr: foundPR,
-      ...(foundTooltip ? { tooltipData: foundTooltip } : {}),
-    });
-  }
-
-  return results;
-}
-
-/**
- * Conditional REST probe against `/repos/{owner}/{repo}/pulls/{number}`.
- * A 304 (Not Modified) response signals the PR has not changed since the
- * stored ETag and does NOT consume primary rate-limit quota for authenticated
- * requests. Returns:
- *   - "unchanged" on 304 (cached ETag still valid)
- *   - "changed" on 200 (new data available; ETag refreshed when present, or
- *     invalidated when the response omits one)
- *   - "unknown" on any other outcome (network error, non-200/304 status, etc.)
- *
- * Callers should treat "unknown" as a cue to fall through to GraphQL —
- * preserving correctness even if the REST path is transiently unavailable.
- */
-async function probePRChange(
-  owner: string,
-  repo: string,
-  prNumber: number,
-  token: string
-): Promise<"changed" | "unchanged" | "unknown"> {
-  const cacheKey = `${owner}/${repo}#${prNumber}`;
-  const cachedETag = prETagCache.get(cacheKey);
-  const url = `https://api.github.com/repos/${owner}/${repo}/pulls/${prNumber}`;
-  const headers: Record<string, string> = {
-    Authorization: `Bearer ${token}`,
-    Accept: "application/vnd.github+json",
-    "X-GitHub-Api-Version": "2022-11-28",
-  };
-  if (cachedETag) {
-    headers["If-None-Match"] = cachedETag;
-  }
-
-  try {
-    const response = await fetch(url, {
-      headers,
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-
-    try {
-      gitHubRateLimitService.update(response.headers, response.status);
-    } catch {
-      // Rate-limit bookkeeping must never break the probe.
-    }
-
-    if (response.status === 304) {
-      return "unchanged";
-    }
-    if (response.status === 200) {
-      const etag = response.headers.get("etag");
-      if (etag) {
-        prETagCache.set(cacheKey, etag);
-      } else {
-        // If the new 200 response carries no ETag, drop any stale cached
-        // validator so the next cycle does not send a now-meaningless
-        // If-None-Match header.
-        prETagCache.delete(cacheKey);
-      }
-      return "changed";
-    }
-    return "unknown";
-  } catch {
-    return "unknown";
-  }
-}
-
-/**
- * Conditional REST probe against `/repos/{owner}/{repo}/pulls?head=...&state=all`
- * used by the discovery path to detect "still no PR for this branch" without
- * paying the GraphQL cost. The `head` filter requires `{owner}:{branch}` —
- * omitting the owner prefix silently bypasses the filter and returns the
- * whole-repo list ETag. Returns:
- *   - "unchanged" on 304, or 200 with an empty array (refreshes ETag)
- *   - "changed"   on 200 with one or more PRs (refreshes ETag; caller falls
- *     through to GraphQL for full enrichment)
- *   - "unknown"   on any other outcome (network error, non-200/304 status,
- *     malformed body) — caller treats as a cue to fall through to GraphQL
- */
-async function probeBranchPRListChange(
-  owner: string,
-  repo: string,
-  branchName: string,
-  token: string
-): Promise<"changed" | "unchanged" | "unknown"> {
-  const cacheKey = `${owner}/${repo}@${branchName}`;
-  const cachedETag = branchListETagCache.get(cacheKey);
-  const headFilter = `${owner}:${encodeURIComponent(branchName)}`;
-  const url = `https://api.github.com/repos/${owner}/${repo}/pulls?head=${headFilter}&state=all`;
-  const headers: Record<string, string> = {
-    Authorization: `Bearer ${token}`,
-    Accept: "application/vnd.github+json",
-    "X-GitHub-Api-Version": "2022-11-28",
-  };
-  if (cachedETag) {
-    headers["If-None-Match"] = cachedETag;
-  }
-
-  try {
-    const response = await fetch(url, {
-      headers,
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-
-    try {
-      gitHubRateLimitService.update(response.headers, response.status);
-    } catch {
-      // Rate-limit bookkeeping must never break the probe.
-    }
-
-    if (response.status === 304) {
-      return "unchanged";
-    }
-    if (response.status === 200) {
-      const etag = response.headers.get("etag");
-      if (etag) {
-        branchListETagCache.set(cacheKey, etag);
-      } else {
-        branchListETagCache.delete(cacheKey);
-      }
-      let body: unknown;
-      try {
-        body = await response.json();
-      } catch {
-        // Malformed body — drop the ETag we just stored (it cannot be trusted
-        // to revalidate against an unparseable response) and fall through.
-        branchListETagCache.delete(cacheKey);
-        return "unknown";
-      }
-      if (!Array.isArray(body)) {
-        // Pathological 200 with a non-array body — drop the cached validator
-        // so the next cycle does not revalidate against an unparseable shape.
-        branchListETagCache.delete(cacheKey);
-        return "unknown";
-      }
-      return body.length === 0 ? "unchanged" : "changed";
-    }
-    return "unknown";
-  } catch {
-    return "unknown";
-  }
-}
-
-/**
- * When every candidate has a `knownPRNumber` (the revalidation path), probe
- * each unique PR via REST conditional requests. If the entire batch is
- * unchanged, the caller can skip the GraphQL query — the dominant per-cycle
- * cost for PR polling — at zero primary quota impact. Duplicate PR numbers
- * across candidates (multiple worktrees pointing at the same PR) are probed
- * only once.
- */
-async function allKnownPRsUnchanged(
-  owner: string,
-  repo: string,
-  candidates: PRCheckCandidate[],
-  token: string
-): Promise<boolean> {
-  const uniquePRNumbers = Array.from(new Set(candidates.map((c) => c.knownPRNumber!)));
-  const probes = await Promise.all(
-    uniquePRNumbers.map((prNumber) => probePRChange(owner, repo, prNumber, token))
-  );
-  return probes.every((result) => result === "unchanged");
-}
-
-export async function batchCheckLinkedPRs(
-  cwd: string,
-  candidates: PRCheckCandidate[]
-): Promise<BatchPRCheckResult> {
-  if (candidates.length === 0) {
-    return { results: new Map() };
-  }
-
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    return { results: new Map(), error: "GitHub token not configured. Set it in Settings." };
-  }
-
-  const context = await getRepoContext(cwd);
-  if (!context) {
-    return { results: new Map(), error: "Not a GitHub repository" };
-  }
-
-  const rateLimitBlock = gitHubRateLimitService.shouldBlockRequest();
-  if (rateLimitBlock.blocked && rateLimitBlock.reason && rateLimitBlock.resumeAt) {
-    return {
-      results: new Map(),
-      error: rateLimitMessage(rateLimitBlock.reason, rateLimitBlock.resumeAt),
-      rateLimit: { kind: rateLimitBlock.reason, resumeAt: rateLimitBlock.resumeAt },
-    };
-  }
-
-  // ETag-gated fast path: if every candidate is a known PR (revalidation),
-  // probe each REST endpoint and skip GraphQL entirely when nothing has
-  // changed. Discovery cycles (any candidate without knownPRNumber) fall
-  // through to the normal GraphQL path.
-  const token = GitHubAuth.getToken();
-  const allRevalidation =
-    token !== undefined && candidates.every((c) => typeof c.knownPRNumber === "number");
-  if (allRevalidation) {
-    const allUnchanged = await allKnownPRsUnchanged(context.owner, context.repo, candidates, token);
-    if (allUnchanged) {
-      return { results: new Map() };
-    }
-    // A probe may itself have observed a 403/429 and updated the rate-limit
-    // service. Re-check before issuing the fallthrough GraphQL request so we
-    // do not burn an attempt while a known pause is already in effect.
-    const postProbeBlock = gitHubRateLimitService.shouldBlockRequest();
-    if (postProbeBlock.blocked && postProbeBlock.reason && postProbeBlock.resumeAt) {
-      return {
-        results: new Map(),
-        error: rateLimitMessage(postProbeBlock.reason, postProbeBlock.resumeAt),
-        rateLimit: { kind: postProbeBlock.reason, resumeAt: postProbeBlock.resumeAt },
-      };
-    }
-  }
-
-  // Discovery-path optimization: for each branch-only candidate (no
-  // knownPRNumber AND no issueNumber, branchName is the sole discovery
-  // signal), run a conditional REST probe against the filtered pulls list.
-  // A 304 or empty 200 means "still no PR for this branch" — that candidate
-  // is dropped from the GraphQL batch entirely. 200-with-results and probe
-  // errors keep the candidate so GraphQL provides full enrichment.
-  //
-  // Candidates with an issueNumber must always pass through to GraphQL: the
-  // PR may be linked via the issue's timeline (CrossReferencedEvent /
-  // ConnectedEvent) on a different head ref than the local branchName, in
-  // which case the branch-list filter would falsely report "no PR."
-  let candidatesForGraphQL = candidates;
-  if (token !== undefined && !allRevalidation) {
-    const probeableIndicesByBranch = new Map<string, number[]>();
-    for (let i = 0; i < candidates.length; i++) {
-      const c = candidates[i];
-      const branch = c.branchName?.trim();
-      if (typeof c.knownPRNumber !== "number" && typeof c.issueNumber !== "number" && branch) {
-        const existing = probeableIndicesByBranch.get(branch);
-        if (existing) {
-          existing.push(i);
-        } else {
-          probeableIndicesByBranch.set(branch, [i]);
-        }
-      }
-    }
-    if (probeableIndicesByBranch.size > 0) {
-      const uniqueBranches = Array.from(probeableIndicesByBranch.keys());
-      const probeResults = await Promise.all(
-        uniqueBranches.map((branch) =>
-          probeBranchPRListChange(context.owner, context.repo, branch, token)
-        )
-      );
-      const skip = new Set<number>();
-      for (let i = 0; i < uniqueBranches.length; i++) {
-        if (probeResults[i] === "unchanged") {
-          for (const idx of probeableIndicesByBranch.get(uniqueBranches[i])!) {
-            skip.add(idx);
-          }
-        }
-      }
-      if (skip.size === candidates.length) {
-        return { results: new Map() };
-      }
-      if (skip.size > 0) {
-        candidatesForGraphQL = candidates.filter((_, idx) => !skip.has(idx));
-      }
-      // A probe may itself have observed a 403/429 and updated the rate-limit
-      // service. Re-check before falling through to GraphQL so we do not burn
-      // an attempt while a known pause is already in effect.
-      const postProbeBlock = gitHubRateLimitService.shouldBlockRequest();
-      if (postProbeBlock.blocked && postProbeBlock.reason && postProbeBlock.resumeAt) {
-        return {
-          results: new Map(),
-          error: rateLimitMessage(postProbeBlock.reason, postProbeBlock.resumeAt),
-          rateLimit: { kind: postProbeBlock.reason, resumeAt: postProbeBlock.resumeAt },
-        };
-      }
-    }
-  }
-
-  try {
-    // Capture timestamp BEFORE the network round-trip so the newest-wins guard
-    // orders writes by request-start time, not response-arrival time. A slow
-    // batch that starts before a fast hover but resolves after must NOT
-    // overwrite the fresher hover write.
-    const requestedAt = Date.now();
-    const query = buildBatchPRQuery(context.owner, context.repo, candidatesForGraphQL);
-    const response = (await client(query, {
-      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-    })) as Record<string, unknown>;
-
-    const results = parseBatchPRResponse(response, candidatesForGraphQL);
-    prewarmPRTooltipCache(context.owner, context.repo, results, requestedAt);
-    return { results };
-  } catch (error) {
-    if (isRepoNotFoundError(error)) {
-      repoContextCache.invalidate(cwd);
-      const freshContext = await getRepoContext(cwd);
-      if (
-        freshContext &&
-        (freshContext.owner !== context.owner || freshContext.repo !== context.repo)
-      ) {
-        try {
-          const retryRequestedAt = Date.now();
-          const retryQuery = buildBatchPRQuery(
-            freshContext.owner,
-            freshContext.repo,
-            candidatesForGraphQL
-          );
-          const retryResponse = (await client(retryQuery, {
-            request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-          })) as Record<string, unknown>;
-          const retryResults = parseBatchPRResponse(retryResponse, candidatesForGraphQL);
-          prewarmPRTooltipCache(
-            freshContext.owner,
-            freshContext.repo,
-            retryResults,
-            retryRequestedAt
-          );
-          return { results: retryResults };
-        } catch (retryError) {
-          return { results: new Map(), error: parseGitHubError(retryError), ...rateLimitMeta() };
-        }
-      }
-    }
-    return { results: new Map(), error: parseGitHubError(error), ...rateLimitMeta() };
-  }
-}
-
-/**
- * Write each batch result's harvested tooltip data into `prTooltipCache`
- * subject to the newest-wins guard. Cache key matches `getPRTooltip`'s
- * format so on-hover IPC calls hit the warm entry instead of issuing a
- * fresh GraphQL request.
- */
-function prewarmPRTooltipCache(
-  owner: string,
-  repo: string,
-  results: Map<string, PRCheckResult>,
-  requestedAt: number
-): void {
-  for (const result of results.values()) {
-    const prNumber = result.pr?.number;
-    const tooltipData = result.tooltipData;
-    if (typeof prNumber !== "number" || !tooltipData) continue;
-    const cacheKey = `${owner}/${repo}:${prNumber}`;
-    const existing = prTooltipWrittenAt.get(cacheKey);
-    if (existing === undefined || requestedAt >= existing) {
-      prTooltipCache.set(cacheKey, tooltipData);
-      prTooltipWrittenAt.set(cacheKey, requestedAt);
-    }
-  }
-}
-
-/**
- * Capture the rate-limit snapshot synchronously after a caller's request
- * fails, so downstream schedulers can route the failure as a rate-limit
- * pause even if a concurrent 2xx clears the singleton before `handleError`
- * reads it. Returns an empty object when no block is active so it spreads
- * cleanly into result objects via rest-spread.
- */
-function rateLimitMeta(): { rateLimit?: { kind: "primary" | "secondary"; resumeAt: number } } {
-  const block = gitHubRateLimitService.shouldBlockRequest();
-  if (block.blocked && block.reason && block.resumeAt) {
-    return { rateLimit: { kind: block.reason, resumeAt: block.resumeAt } };
-  }
-  return {};
-}
-
-export async function getRepoUrl(cwd: string): Promise<string | null> {
-  const context = await getRepoContext(cwd);
-  if (!context) return null;
-  return `https://github.com/${context.owner}/${context.repo}`;
-}
-
-export async function getIssueUrl(cwd: string, issueNumber: number): Promise<string | null> {
-  const repoUrl = await getRepoUrl(cwd);
-  if (!repoUrl) return null;
-  return `${repoUrl}/issues/${issueNumber}`;
-}
-
-export interface AssignIssueResult {
-  username: string;
-  avatarUrl: string;
-}
-
-export async function assignIssue(
-  cwd: string,
-  issueNumber: number,
-  username: string
-): Promise<AssignIssueResult> {
-  const token = getGitHubToken();
-  if (!token) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-
-  return withRepoContextRetry(cwd, async (context) => {
-    const url = `https://api.github.com/repos/${context.owner}/${context.repo}/issues/${issueNumber}/assignees`;
-
-    try {
-      const response = await fetch(url, {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${token}`,
-          Accept: "application/vnd.github+json",
-          "X-GitHub-Api-Version": "2022-11-28",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ assignees: [username] }),
-        signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-      });
-
-      if (!response.ok) {
-        if (response.status === 401) {
-          throw new Error("Invalid GitHub token. Please update in Settings.");
-        }
-        if (response.status === 403) {
-          throw new Error("Token lacks required permissions. Required scopes: repo, read:org");
-        }
-        if (response.status === 404) {
-          throw new Error("Issue not found or you don't have access to this repository");
-        }
-        if (response.status === 422) {
-          throw new Error(`Cannot assign user "${username}" - they may not be a collaborator`);
-        }
-        throw new Error(`GitHub API error: ${response.statusText}`);
-      }
-
-      const data = (await response.json()) as {
-        assignees?: Array<{ login?: string; avatar_url?: string }>;
-      };
-
-      if (!Array.isArray(data.assignees)) {
-        throw new Error("Invalid GitHub API response: assignees field missing or malformed");
-      }
-
-      const assignee = data.assignees.find(
-        (a) => a.login?.toLowerCase() === username.toLowerCase()
-      );
-      if (!assignee?.login) {
-        throw new Error(`Assignment succeeded but user "${username}" not found in response`);
-      }
-
-      const assigneeData = {
-        login: assignee.login,
-        avatarUrl: assignee.avatar_url ?? "",
-      };
-
-      updateIssueAssigneeInCache(context.owner, context.repo, issueNumber, assigneeData);
-
-      return { username: assigneeData.login, avatarUrl: assigneeData.avatarUrl };
-    } catch (error) {
-      throw new Error(parseGitHubError(error));
-    }
-  });
-}
-
-function updateIssueAssigneeInCache(
-  owner: string,
-  repo: string,
-  issueNumber: number,
-  assignee: { login: string; avatarUrl: string }
-): void {
-  // The cache key format is: type:owner/repo:state:search:cursor
-  // We need to iterate through all cache entries and update any that contain this issue
-  const cachePrefix = `issue:${owner}/${repo}:`;
-
-  // Collect all updates first to avoid modifying cache during iteration
-  const updates: Array<{ key: string; value: GitHubListResponse<GitHubIssue> }> = [];
-
-  issueListCache.forEach((value, key) => {
-    if (!key.startsWith(cachePrefix)) return;
-
-    const issueIndex = value.items.findIndex((issue) => issue.number === issueNumber);
-    if (issueIndex === -1) return;
-
-    // Check if assignee already exists
-    const existingAssignees = value.items[issueIndex].assignees;
-    const existingIndex = existingAssignees.findIndex(
-      (a) => a.login.toLowerCase() === assignee.login.toLowerCase()
-    );
-
-    let updatedAssignees: GitHubUser[];
-    if (existingIndex !== -1) {
-      // Update existing assignee if avatarUrl changed
-      if (existingAssignees[existingIndex].avatarUrl !== assignee.avatarUrl) {
-        updatedAssignees = [...existingAssignees];
-        updatedAssignees[existingIndex] = assignee;
-      } else {
-        return;
-      }
-    } else {
-      // Add new assignee
-      updatedAssignees = [...existingAssignees, assignee];
-    }
-
-    // Create updated items array
-    const updatedItems = [...value.items];
-    updatedItems[issueIndex] = {
-      ...updatedItems[issueIndex],
-      assignees: updatedAssignees,
-    };
-
-    updates.push({
-      key,
-      value: {
-        ...value,
-        items: updatedItems,
-      },
-    });
-  });
-
-  // Apply all updates after iteration completes
-  for (const update of updates) {
-    issueListCache.set(update.key, update.value);
-  }
-}
-
-function rateLimitMessage(kind: "primary" | "secondary", resumeAt: number): string {
-  const seconds = Math.max(0, Math.ceil((resumeAt - Date.now()) / 1000));
-  const human = formatCountdown(seconds);
-  if (kind === "secondary") {
-    return `GitHub secondary rate limit triggered. Resuming in ${human}.`;
-  }
-  return `GitHub rate limit exceeded. Resets in ${human}.`;
-}
-
-function formatCountdown(totalSeconds: number): string {
-  if (totalSeconds <= 0) return "a moment";
-  if (totalSeconds < 60) return `${totalSeconds}s`;
-  const minutes = Math.floor(totalSeconds / 60);
-  const seconds = totalSeconds % 60;
-  if (minutes < 60) {
-    return seconds > 0 ? `${minutes}m ${seconds}s` : `${minutes}m`;
-  }
-  const hours = Math.floor(minutes / 60);
-  const remMinutes = minutes % 60;
-  return remMinutes > 0 ? `${hours}h ${remMinutes}m` : `${hours}h`;
-}
-
-export function parseGitHubError(error: unknown): string {
-  if (error instanceof GitHubRateLimitError) {
-    return rateLimitMessage(error.kind, error.resumeAt);
-  }
-  const message = formatErrorMessage(error, "GitHub request failed");
-  const isTimeout = error instanceof Error && error.name === "TimeoutError";
-
-  if (
-    message === "GitHub token not configured. Set it in Settings." ||
-    message === "Invalid GitHub token. Please update in Settings." ||
-    message === "Token lacks required permissions. Required scopes: repo, read:org" ||
-    message === "Issue not found or you don't have access to this repository" ||
-    message.startsWith("Cannot assign user ") ||
-    message.startsWith("Assignment succeeded but user ") ||
-    message.startsWith("Invalid GitHub API response:") ||
-    message === "Cannot reach GitHub. Check your internet connection."
-  ) {
-    return message;
-  }
-
-  const blockState = gitHubRateLimitService.shouldBlockRequest();
-  if (blockState.blocked && blockState.reason && blockState.resumeAt) {
-    return rateLimitMessage(blockState.reason, blockState.resumeAt);
-  }
-
-  if (message.includes("rate limit") || message.includes("API rate limit")) {
-    return "GitHub rate limit exceeded. Try again in a few minutes.";
-  }
-
-  if (message.includes("401") || message.includes("Bad credentials")) {
-    return "Invalid GitHub token. Please update in Settings.";
-  }
-
-  // Check SAML/SSO *before* the generic 403 branch — GitHub reports SSO
-  // enforcement failures as "Resource protected by organization SAML
-  // enforcement... (403)", so the substring "403" would otherwise swallow
-  // them into the generic permissions message and we'd lose the captured
-  // re-authorization URL.
-  if (message.includes("SAML") || message.includes("SSO")) {
-    // `rateLimitAwareFetch` captures the `X-GitHub-SSO: required; url=...`
-    // header passively. Surface the exact re-authorization URL when we have
-    // one so the user can re-authorize in a single click instead of hunting
-    // through github.com.
-    const ssoUrl = getLastAuthMetadata()?.ssoUrl;
-    if (ssoUrl) {
-      return `SSO authorization required. Re-authorize at: ${ssoUrl}`;
-    }
-    return "SSO authorization required. Re-authorize at github.com.";
-  }
-
-  if (message.includes("403")) {
-    return "Token lacks required permissions. Required scopes: repo, read:org";
-  }
-
-  if (message.includes("404") || message.includes("Could not resolve")) {
-    return "Repository not found or token lacks access.";
-  }
-
-  if (
-    isTimeout ||
-    message.includes("ENOTFOUND") ||
-    message.includes("ETIMEDOUT") ||
-    message.includes("ECONNREFUSED") ||
-    message.includes("ECONNRESET") ||
-    message.includes("EAI_AGAIN") ||
-    message.includes("network") ||
-    message.includes("fetch failed") ||
-    message.includes("timed out")
-  ) {
-    return "Cannot reach GitHub. Check your internet connection.";
-  }
-
-  return `GitHub API error: ${message}`;
-}
-
-export function clearGitHubCaches(): void {
-  repoContextCache.clear();
-  repoStatsCache.clear();
-  projectHealthCache.clear();
-  issueListCache.clear();
-  prListCache.clear();
-  issueTooltipCache.clear();
-  prTooltipCache.clear();
-  prTooltipWrittenAt.clear();
-  prETagCache.clear();
-  branchListETagCache.clear();
-  prRequiredStatusCache.clear();
-  // Token rotation invalidates persisted first-page data — a different
-  // identity may have access to a different repo set.
-  GitHubFirstPageCache.getInstance().clear();
-}
-
-export function clearPRCaches(): void {
-  prListCache.clear();
-  prTooltipCache.clear();
-  prTooltipWrittenAt.clear();
-  prRequiredStatusCache.clear();
-}
-
-function buildListCacheKey(
-  type: "issue" | "pr",
-  owner: string,
-  repo: string,
-  state: string,
-  search: string,
-  sortOrder: string,
-  cursor: string
-): string {
-  return `${type}:${owner}/${repo}:${state}:${search}:${sortOrder}:${cursor}`;
-}
-
-function mapIssueStates(state?: string): string[] {
-  if (!state || state === "open") return ["OPEN"];
-  if (state === "closed") return ["CLOSED"];
-  if (state === "all") return ["OPEN", "CLOSED"];
-  return ["OPEN"];
-}
-
-function updateRepoStatsCount(cacheKey: string, type: "issue" | "pr", count: number): void {
-  const cached = repoStatsCache.get(cacheKey);
-  if (cached) {
-    const updated: RepoStats = {
-      ...cached,
-      lastUpdated: Date.now(),
-    };
-    if (type === "issue") {
-      updated.issueCount = count;
-    } else {
-      updated.prCount = count;
-    }
-    repoStatsCache.set(cacheKey, updated);
-  } else {
-    const persistentCache = GitHubStatsCache.getInstance();
-    const diskCached = persistentCache.get(cacheKey);
-
-    const newStats: RepoStats = {
-      issueCount: type === "issue" ? count : (diskCached?.issueCount ?? 0),
-      prCount: type === "pr" ? count : (diskCached?.prCount ?? 0),
-      lastUpdated: Date.now(),
-    };
-    repoStatsCache.set(cacheKey, newStats);
-  }
-}
-
-function mapPRStates(state?: string): string[] {
-  if (!state || state === "open") return ["OPEN"];
-  if (state === "closed") return ["CLOSED"];
-  if (state === "merged") return ["MERGED"];
-  if (state === "all") return ["OPEN", "CLOSED", "MERGED"];
-  return ["OPEN"];
-}
-
-function extractLinkedPR(
-  timelineItems:
-    | {
-        nodes?: Array<{
-          // CROSS_REFERENCED_EVENT uses 'source'
-          source?: { number?: number; state?: string; merged?: boolean; url?: string };
-          // CONNECTED_EVENT uses 'subject'
-          subject?: { number?: number; state?: string; merged?: boolean; url?: string };
-        }>;
-      }
-    | undefined
-): LinkedPRInfo | undefined {
-  if (!timelineItems?.nodes) return undefined;
-
-  const prs: Array<{ number: number; state: "OPEN" | "CLOSED" | "MERGED"; url: string }> = [];
-  const seenPRNumbers = new Set<number>();
-
-  for (const node of timelineItems.nodes) {
-    // Handle both CROSS_REFERENCED_EVENT (source) and CONNECTED_EVENT (subject)
-    const prData = node?.source ?? node?.subject;
-    if (prData?.number && prData?.url && !seenPRNumbers.has(prData.number)) {
-      seenPRNumbers.add(prData.number);
-      const state: "OPEN" | "CLOSED" | "MERGED" = prData.merged
-        ? "MERGED"
-        : (prData.state?.toUpperCase() as "OPEN" | "CLOSED") || "OPEN";
-      prs.push({ number: prData.number, state, url: prData.url });
-    }
-  }
-
-  if (prs.length === 0) return undefined;
-
-  const openPRs = prs.filter((pr) => pr.state === "OPEN");
-  const mergedPRs = prs.filter((pr) => pr.state === "MERGED");
-  const closedPRs = prs.filter((pr) => pr.state === "CLOSED");
-
-  if (openPRs.length > 0) return openPRs[openPRs.length - 1];
-  if (mergedPRs.length > 0) return mergedPRs[mergedPRs.length - 1];
-  if (closedPRs.length > 0) return closedPRs[closedPRs.length - 1];
-
-  return undefined;
-}
-
-function parseIssueNode(node: Record<string, unknown>): GitHubIssue {
-  const author = node.author as { login?: string; avatarUrl?: string } | null;
-  const assigneesData = node.assignees as { nodes?: Array<{ login?: string; avatarUrl?: string }> };
-  const commentsData = node.comments as { totalCount?: number };
-  const labelsData = node.labels as { nodes?: Array<{ name?: string; color?: string }> };
-  const timelineItems = node.timelineItems as
-    | {
-        nodes?: Array<{
-          source?: { number?: number; state?: string; merged?: boolean; url?: string };
-        }>;
-      }
-    | undefined;
-
-  const linkedPR = extractLinkedPR(timelineItems);
-
-  return {
-    number: node.number as number,
-    title: node.title as string,
-    url: node.url as string,
-    state: node.state as "OPEN" | "CLOSED",
-    updatedAt: node.updatedAt as string,
-    author: {
-      login: author?.login ?? "unknown",
-      avatarUrl: author?.avatarUrl ?? "",
-    },
-    assignees: (assigneesData?.nodes ?? []).filter(Boolean).map((a) => ({
-      login: a.login ?? "unknown",
-      avatarUrl: a.avatarUrl ?? "",
-    })),
-    commentCount: commentsData?.totalCount ?? 0,
-    labels: (labelsData?.nodes ?? []).filter(Boolean).map((l) => ({
-      name: l.name ?? "",
-      color: l.color ?? "",
-    })),
-    linkedPR,
-  };
-}
-
-function parsePRNode(node: Record<string, unknown>): GitHubPR {
-  const author = node.author as { login?: string; avatarUrl?: string } | null;
-  const reviewsData = node.reviews as { totalCount?: number };
-  const commentsData = node.comments as { totalCount?: number };
-  const merged = node.merged as boolean;
-  const rawState = node.state as string;
-  const headRepo = node.headRepository as { nameWithOwner?: string } | null;
-  const baseRepo = node.baseRepository as { nameWithOwner?: string } | null;
-
-  let state: "OPEN" | "CLOSED" | "MERGED" = rawState as "OPEN" | "CLOSED" | "MERGED";
-  if (merged) {
-    state = "MERGED";
-  }
-
-  const headName = headRepo?.nameWithOwner;
-  const baseName = baseRepo?.nameWithOwner;
-  const isFork = headName && baseName ? headName !== baseName : undefined;
-
-  const commitsData = node.commits as
-    | { nodes?: Array<{ commit?: { statusCheckRollup?: { state?: string } | null } | null }> }
-    | undefined;
-  const ciStatus = commitsData?.nodes?.[0]?.commit?.statusCheckRollup?.state as
-    | GitHubPRCIStatus
-    | undefined;
-
-  return {
-    number: node.number as number,
-    title: node.title as string,
-    url: node.url as string,
-    state,
-    isDraft: (node.isDraft as boolean) ?? false,
-    updatedAt: node.updatedAt as string,
-    author: {
-      login: author?.login ?? "unknown",
-      avatarUrl: author?.avatarUrl ?? "",
-    },
-    reviewCount: reviewsData?.totalCount,
-    commentCount: commentsData?.totalCount,
-    headRefName: (node.headRefName as string) || undefined,
-    isFork: isFork ?? undefined,
-    ciStatus,
-  };
-}
-
-export async function listIssues(
-  options: GitHubListOptions
-): Promise<GitHubListResponse<GitHubIssue>> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-
-  const resolvedSortOrder = options.sortOrder ?? "created";
-  const orderBy = {
-    field: resolvedSortOrder === "updated" ? "UPDATED_AT" : "CREATED_AT",
-    direction: "DESC",
-  };
-
-  return withRepoContextRetry(options.cwd, async (context) => {
-    const cacheKey = buildListCacheKey(
-      "issue",
-      context.owner,
-      context.repo,
-      options.state ?? "open",
-      options.search ?? "",
-      resolvedSortOrder,
-      options.cursor ?? ""
-    );
-
-    if (!options.search && !options.bypassCache) {
-      const cached = issueListCache.get(cacheKey);
-      if (cached) return cached;
-    }
-
-    try {
-      let result: GitHubListResponse<GitHubIssue>;
-
-      if (options.search) {
-        const stateFilter =
-          options.state === "closed" ? "is:closed" : options.state === "all" ? "" : "is:open";
-        const sortQualifier =
-          resolvedSortOrder === "updated" ? "sort:updated-desc" : "sort:created-desc";
-        const searchQuery =
-          `repo:${context.owner}/${context.repo} is:issue ${stateFilter} ${sortQualifier} ${options.search}`.trim();
-
-        const response = (await client(SEARCH_QUERY, {
-          searchQuery,
-          type: "ISSUE",
-          cursor: options.cursor,
-          limit: 20,
-          request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-        })) as GraphQlQueryResponseData;
-
-        const search = response?.search;
-        const nodes = (search?.nodes ?? []) as Array<Record<string, unknown>>;
-
-        result = {
-          items: nodes.filter(Boolean).map(parseIssueNode),
-          pageInfo: {
-            hasNextPage: search?.pageInfo?.hasNextPage ?? false,
-            endCursor: search?.pageInfo?.endCursor ?? null,
-          },
-        };
-      } else {
-        const states = mapIssueStates(options.state);
-
-        const response = (await client(LIST_ISSUES_QUERY, {
-          owner: context.owner,
-          repo: context.repo,
-          states,
-          cursor: options.cursor,
-          limit: 20,
-          orderBy,
-          request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-        })) as GraphQlQueryResponseData;
-
-        const issues = response?.repository?.issues;
-        const nodes = (issues?.nodes ?? []) as Array<Record<string, unknown>>;
-        const totalCount = (issues?.totalCount as number) ?? undefined;
-
-        result = {
-          items: nodes.filter(Boolean).map(parseIssueNode),
-          pageInfo: {
-            hasNextPage: issues?.pageInfo?.hasNextPage ?? false,
-            endCursor: issues?.pageInfo?.endCursor ?? null,
-          },
-        };
-
-        issueListCache.set(cacheKey, result);
-
-        if (
-          (!options.state || options.state === "open") &&
-          !options.cursor &&
-          totalCount !== undefined
-        ) {
-          const statsCacheKey = `${context.owner}/${context.repo}`;
-          updateRepoStatsCount(statsCacheKey, "issue", totalCount);
-
-          const memoryStats = repoStatsCache.get(statsCacheKey);
-          if (memoryStats && memoryStats.issueCount > 0 && memoryStats.prCount > 0) {
-            const persistentCache = GitHubStatsCache.getInstance();
-            persistentCache.set(
-              statsCacheKey,
-              {
-                issueCount: memoryStats.issueCount,
-                prCount: memoryStats.prCount,
-              },
-              options.cwd
-            );
-          }
-        }
-      }
-
-      return result;
-    } catch (error) {
-      throw new Error(parseGitHubError(error));
-    }
-  });
-}
-
-/**
- * For each open PR with a raw statusCheckRollup state and no cached ciSummary,
- * fetch per-PR contexts with `isRequired` and derive an effective CI status
- * based on required checks only. Re-runs on every list query so a SUCCESS PR
- * that flips back to PENDING/FAILURE on a re-push refreshes within the
- * `prRequiredStatusCache` TTL. Best-effort: on any error, returns input unchanged.
- */
-async function enrichPRsWithRequiredStatus(
-  context: RepoContext,
-  prs: GitHubPR[],
-  client: NonNullable<ReturnType<typeof GitHubAuth.createClient>>,
-  bypassCache = false
-): Promise<GitHubPR[]> {
-  const candidates = prs.filter(
-    (pr) => pr.state === "OPEN" && pr.ciStatus !== undefined && pr.ciSummary === undefined
-  );
-  if (candidates.length === 0) return prs;
-
-  const cacheKeyFor = (n: number) => `${context.owner}/${context.repo}:${n}`;
-  const numbersToFetch: number[] = [];
-  const cached = new Map<number, PRRequiredStatusEntry>();
-
-  for (const pr of candidates) {
-    const hit = bypassCache ? undefined : prRequiredStatusCache.get(cacheKeyFor(pr.number));
-    if (hit) {
-      cached.set(pr.number, hit);
-    } else {
-      numbersToFetch.push(pr.number);
-    }
-  }
-
-  let fetched = new Map<number, PRRequiredStatusEntry>();
-  if (numbersToFetch.length > 0) {
-    try {
-      const query = buildBatchRequiredChecksQuery(context.owner, context.repo, numbersToFetch);
-      if (query) {
-        const response = (await client(query, {
-          request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-        })) as Record<string, unknown>;
-        fetched = parseBatchRequiredChecksResponse(response, numbersToFetch);
-        for (const [num, entry] of fetched) {
-          prRequiredStatusCache.set(cacheKeyFor(num), entry);
-        }
-      }
-    } catch {
-      // Best-effort — fall through; PRs keep their raw ciStatus and no ciSummary.
-    }
-  }
-
-  return prs.map((pr) => {
-    const entry = cached.get(pr.number) ?? fetched.get(pr.number);
-    if (!entry) return pr;
-    return {
-      ...pr,
-      ciStatus: entry.ciStatus ?? pr.ciStatus,
-      ciSummary: entry.ciSummary,
-    };
-  });
-}
-
-function parseBatchRequiredChecksResponse(
-  data: Record<string, unknown>,
-  prNumbers: number[]
-): Map<number, PRRequiredStatusEntry> {
-  const out = new Map<number, PRRequiredStatusEntry>();
-  for (const num of prNumbers) {
-    const alias = `pr_${num}`;
-    const repo = data?.[alias] as
-      | {
-          pullRequest?: {
-            commits?: {
-              nodes?: Array<{
-                commit?: {
-                  statusCheckRollup?: {
-                    state?: string | null;
-                    contexts?: {
-                      pageInfo?: { hasNextPage?: boolean };
-                      nodes?: RollupContextNode[];
-                    };
-                  } | null;
-                } | null;
-              }>;
-            };
-          };
-        }
-      | undefined;
-    const rollup = repo?.pullRequest?.commits?.nodes?.[0]?.commit?.statusCheckRollup;
-    if (!rollup) continue;
-    const derived = deriveRequiredCIStatus(
-      rollup.contexts?.nodes,
-      rollup.contexts?.pageInfo?.hasNextPage ?? false,
-      rollup.state
-    );
-    out.set(num, { ciStatus: derived.ciStatus, ciSummary: derived.ciSummary });
-  }
-  return out;
-}
-
-export async function listPullRequests(
-  options: GitHubListOptions
-): Promise<GitHubListResponse<GitHubPR>> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-
-  const resolvedSortOrder = options.sortOrder ?? "created";
-  const orderBy = {
-    field: resolvedSortOrder === "updated" ? "UPDATED_AT" : "CREATED_AT",
-    direction: "DESC",
-  };
-
-  return withRepoContextRetry(options.cwd, async (context) => {
-    const cacheKey = buildListCacheKey(
-      "pr",
-      context.owner,
-      context.repo,
-      options.state ?? "open",
-      options.search ?? "",
-      resolvedSortOrder,
-      options.cursor ?? ""
-    );
-
-    if (!options.search && !options.bypassCache) {
-      const cached = prListCache.get(cacheKey);
-      if (cached) return cached;
-    }
-
-    try {
-      let result: GitHubListResponse<GitHubPR>;
-
-      if (options.search) {
-        let stateFilter = "";
-        if (options.state === "open") stateFilter = "is:open";
-        else if (options.state === "closed") stateFilter = "is:closed is:unmerged";
-        else if (options.state === "merged") stateFilter = "is:merged";
-
-        const sortQualifier =
-          resolvedSortOrder === "updated" ? "sort:updated-desc" : "sort:created-desc";
-        const searchQuery =
-          `repo:${context.owner}/${context.repo} is:pr ${stateFilter} ${sortQualifier} ${options.search}`.trim();
-
-        const response = (await client(SEARCH_QUERY, {
-          searchQuery,
-          type: "ISSUE",
-          cursor: options.cursor,
-          limit: 20,
-          request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-        })) as GraphQlQueryResponseData;
-
-        const search = response?.search;
-        const nodes = (search?.nodes ?? []) as Array<Record<string, unknown>>;
-
-        const parsedItems = nodes.filter(Boolean).map(parsePRNode);
-        const enrichedItems = await enrichPRsWithRequiredStatus(
-          context,
-          parsedItems,
-          client,
-          options.bypassCache ?? false
-        );
-        result = {
-          items: enrichedItems,
-          pageInfo: {
-            hasNextPage: search?.pageInfo?.hasNextPage ?? false,
-            endCursor: search?.pageInfo?.endCursor ?? null,
-          },
-        };
-      } else {
-        const states = mapPRStates(options.state);
-
-        const response = (await client(LIST_PRS_QUERY, {
-          owner: context.owner,
-          repo: context.repo,
-          states,
-          cursor: options.cursor,
-          limit: 20,
-          orderBy,
-          request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-        })) as GraphQlQueryResponseData;
-
-        if (!response?.repository) {
-          throw new Error("Repository not found or token lacks access.");
-        }
-
-        const pullRequests = response?.repository?.pullRequests;
-        const nodes = (pullRequests?.nodes ?? []) as Array<Record<string, unknown>>;
-        const totalCount = (pullRequests?.totalCount as number) ?? undefined;
-
-        const parsedItems = nodes.filter(Boolean).map(parsePRNode);
-        const enrichedItems = await enrichPRsWithRequiredStatus(
-          context,
-          parsedItems,
-          client,
-          options.bypassCache ?? false
-        );
-        result = {
-          items: enrichedItems,
-          pageInfo: {
-            hasNextPage: pullRequests?.pageInfo?.hasNextPage ?? false,
-            endCursor: pullRequests?.pageInfo?.endCursor ?? null,
-          },
-        };
-
-        prListCache.set(cacheKey, result);
-
-        if (
-          (!options.state || options.state === "open") &&
-          !options.cursor &&
-          totalCount !== undefined
-        ) {
-          const statsCacheKey = `${context.owner}/${context.repo}`;
-          updateRepoStatsCount(statsCacheKey, "pr", totalCount);
-
-          const memoryStats = repoStatsCache.get(statsCacheKey);
-          if (memoryStats && memoryStats.issueCount > 0 && memoryStats.prCount > 0) {
-            const persistentCache = GitHubStatsCache.getInstance();
-            persistentCache.set(
-              statsCacheKey,
-              {
-                issueCount: memoryStats.issueCount,
-                prCount: memoryStats.prCount,
-              },
-              options.cwd
-            );
-          }
-        }
-      }
-
-      return result;
-    } catch (error) {
-      throw new Error(parseGitHubError(error));
-    }
-  });
-}
-
-function truncateBody(body: string | null | undefined, maxLength = 150): string {
-  if (!body) return "";
-  const cleaned = body.replace(/\r?\n/g, " ").trim();
-  if (cleaned.length <= maxLength) return cleaned;
-  return cleaned.slice(0, maxLength).trim() + "…";
-}
-
-export async function getIssueTooltip(
-  cwd: string,
-  issueNumber: number
-): Promise<IssueTooltipData | null> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    return null;
-  }
-
-  try {
-    return await withRepoContextRetry(cwd, async (context) => {
-      const cacheKey = `${context.owner}/${context.repo}:${issueNumber}`;
-      const cached = issueTooltipCache.get(cacheKey);
-      if (cached) {
-        return cached;
-      }
-
-      const response = (await client(GET_ISSUE_QUERY, {
-        owner: context.owner,
-        repo: context.repo,
-        number: issueNumber,
-        request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-      })) as GraphQlQueryResponseData;
-
-      const issue = response?.repository?.issue;
-      if (!issue) {
-        return null;
-      }
-
-      const author = issue.author as { login?: string; avatarUrl?: string } | null;
-      const assigneesData = issue.assignees as {
-        nodes?: Array<{ login?: string; avatarUrl?: string }>;
-      };
-      const labelsData = issue.labels as { nodes?: Array<{ name?: string; color?: string }> };
-
-      const tooltipData: IssueTooltipData = {
-        number: issue.number as number,
-        title: issue.title as string,
-        bodyExcerpt: truncateBody(issue.bodyText as string | null),
-        state: issue.state as "OPEN" | "CLOSED",
-        createdAt: issue.createdAt as string,
-        author: {
-          login: author?.login ?? "unknown",
-          avatarUrl: author?.avatarUrl ?? "",
-        },
-        assignees: (assigneesData?.nodes ?? []).filter(Boolean).map((a) => ({
-          login: a.login ?? "unknown",
-          avatarUrl: a.avatarUrl ?? "",
-        })),
-        labels: (labelsData?.nodes ?? []).filter(Boolean).map((l) => ({
-          name: l.name ?? "",
-          color: l.color ?? "",
-        })),
-      };
-
-      issueTooltipCache.set(cacheKey, tooltipData);
-      return tooltipData;
-    });
-  } catch {
-    return null;
-  }
-}
-
-export async function getPRTooltip(cwd: string, prNumber: number): Promise<PRTooltipData | null> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    return null;
-  }
-
-  try {
-    return await withRepoContextRetry(cwd, async (context) => {
-      const cacheKey = `${context.owner}/${context.repo}:${prNumber}`;
-      const cached = prTooltipCache.get(cacheKey);
-      if (cached) {
-        return cached;
-      }
-
-      // Capture timestamp BEFORE the network round-trip so the newest-wins
-      // guard orders writes by request-start time. A poll batch starting
-      // later would still overwrite if we used response-arrival time here.
-      const requestedAt = Date.now();
-      const response = (await client(GET_PR_QUERY, {
-        owner: context.owner,
-        repo: context.repo,
-        number: prNumber,
-        request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-      })) as GraphQlQueryResponseData;
-
-      const pr = response?.repository?.pullRequest;
-      if (!pr) {
-        return null;
-      }
-
-      const author = pr.author as { login?: string; avatarUrl?: string } | null;
-      const assigneesData = pr.assignees as {
-        nodes?: Array<{ login?: string; avatarUrl?: string }>;
-      };
-      const labelsData = pr.labels as { nodes?: Array<{ name?: string; color?: string }> };
-      const merged = pr.merged as boolean;
-      const rawState = pr.state as string;
-
-      let state: "OPEN" | "CLOSED" | "MERGED" = rawState as "OPEN" | "CLOSED" | "MERGED";
-      if (merged) {
-        state = "MERGED";
-      }
-
-      const tooltipData: PRTooltipData = {
-        number: pr.number as number,
-        title: pr.title as string,
-        bodyExcerpt: truncateBody(pr.bodyText as string | null),
-        state,
-        isDraft: (pr.isDraft as boolean) ?? false,
-        createdAt: pr.createdAt as string,
-        author: {
-          login: author?.login ?? "unknown",
-          avatarUrl: author?.avatarUrl ?? "",
-        },
-        assignees: (assigneesData?.nodes ?? []).filter(Boolean).map((a) => ({
-          login: a.login ?? "unknown",
-          avatarUrl: a.avatarUrl ?? "",
-        })),
-        labels: (labelsData?.nodes ?? []).filter(Boolean).map((l) => ({
-          name: l.name ?? "",
-          color: l.color ?? "",
-        })),
-      };
-
-      const existing = prTooltipWrittenAt.get(cacheKey);
-      if (existing === undefined || requestedAt >= existing) {
-        prTooltipCache.set(cacheKey, tooltipData);
-        prTooltipWrittenAt.set(cacheKey, requestedAt);
-      }
-      return tooltipData;
-    });
-  } catch {
-    return null;
-  }
-}
-
-export async function getIssueByNumber(
-  cwd: string,
-  issueNumber: number
-): Promise<GitHubIssue | null> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-
-  try {
-    return await withRepoContextRetry(cwd, async (context) => {
-      const response = (await client(GET_ISSUE_QUERY, {
-        owner: context.owner,
-        repo: context.repo,
-        number: issueNumber,
-        request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-      })) as GraphQlQueryResponseData;
-
-      const issue = response?.repository?.issue;
-      if (!issue) {
-        return null;
-      }
-
-      return parseIssueNode(issue as Record<string, unknown>);
-    });
-  } catch (error) {
-    const message = formatErrorMessage(error, "Failed to fetch GitHub issue");
-    if (message === "Not a GitHub repository") {
-      throw error;
-    }
-    if (message.includes("Could not resolve to") || message.includes("Could not resolve")) {
-      return null;
-    }
-    throw new Error(parseGitHubError(error));
-  }
-}
-
-export async function getPRByNumber(cwd: string, prNumber: number): Promise<GitHubPR | null> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-
-  try {
-    return await withRepoContextRetry(cwd, async (context) => {
-      const response = (await client(GET_PR_QUERY, {
-        owner: context.owner,
-        repo: context.repo,
-        number: prNumber,
-        request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-      })) as GraphQlQueryResponseData;
-
-      const pr = response?.repository?.pullRequest;
-      if (!pr) {
-        return null;
-      }
-
-      return parsePRNode(pr as Record<string, unknown>);
-    });
-  } catch (error) {
-    const message = formatErrorMessage(error, "Failed to fetch GitHub pull request");
-    if (message === "Not a GitHub repository") {
-      throw error;
-    }
-    if (message.includes("Could not resolve to") || message.includes("Could not resolve")) {
-      return null;
-    }
-    throw new Error(parseGitHubError(error));
-  }
-}
+export type { RepoStatsAndPageResult, AssignIssueResult } from "./github/index.js";
